requirements.txt
flask==3.0.0
pymongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
flask-session==0.5.0

